folder structure moved to DDD pattern
