+++ backend/src/api-gateway/modules/qr-code/qr-code.controller.ts
@@ -1,11 +1,11 @@
 import { Body, Controller, Post, Res } from '@nestjs/common';
 import { QrCodeService } from './qr-code.service';
 import { Response } from 'express';
- 
+
 @Controller('qr-code')
 export class QrCodeController {
   constructor(private readonly qrCodeService: QrCodeService) {}
- 
+
   @Post('generate')
   async generateQRCode(@Body() body: { type: string; data: any }) {
     const { type, data } = body;
@@ -13,7 +13,7 @@ export class QrCodeController {
       qrCode: await this.qrCodeService.generateQRCode(type, data),
     };
   }
- 
+
   @Post('generate-png')
   async generateQRCodePNG(@Body() body: { type: string; data: any }, @Res() res: Response) {
     const { type, data } = body;
@@ -21,7 +21,7 @@ export class QrCodeController {
     res.setHeader('Content-Type', 'image/png');
     res.send(buffer);
   }
- 
+
   @Post('generate-svg')
   async generateQRCodeSVG(@Body() body: { type: string; data: any }, @Res() res: Response) {
     const { type, data } = body;
@@ -29,7 +29,7 @@ export class QrCodeController {
     res.setHeader('Content-Type', 'image/svg+xml');
     res.send(svg);
   }
- 
+
   @Post('generate-pdf')
   async generateQRCodePDF(@Body() body: { type: string; data: any }, @Res() res: Response) {
     const { type, data } = body;
