+++ backend/src/api-gateway/modules/qr-code/qr-code.module.ts
@@ -1,9 +1,10 @@
 import { Module } from '@nestjs/common';
 import { QrCodeService } from './qr-code.service';
 import { QrCodeController } from './qr-code.controller';
+import { QrCodeRepository } from 'src/domain/repositories/qr-code.repository';
 
 @Module({
-  providers: [QrCodeService],
   controllers: [QrCodeController],
+  providers: [QrCodeService, QrCodeRepository],
 })
 export class QrCodeModule {}
