+++ backend/src/api-gateway/qr-code.gateway.ts
@@ -1,0 +1,7 @@
+import { Module } from '@nestjs/common';
+import { QrCodeModule } from './modules/qr-code/qr-code.module';
+
+@Module({
+  imports: [QrCodeModule],
+})
+export class QrCodeGatewayModule {}
