+++ backend/src/app.module.ts
@@ -1,12 +1,7 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
-import { QrCodeService } from './qr-code/qr-code.service';
-import { QrCodeController } from './qr-code/qr-code.controller';
+import { QrCodeGatewayModule } from './api-gateway/qr-code.gateway';
 
 @Module({
-  imports: [],
-  controllers: [AppController, QrCodeController],
-  providers: [AppService, QrCodeService],
+  imports: [QrCodeGatewayModule],
 })
 export class AppModule {}
