+++ backend/src/domain/entities/qr-code.entity.ts
@@ -1,0 +1,7 @@
+export class QrCode {
+    constructor(
+      public readonly type: string,
+      public readonly data: any,
+    ) {}
+  }
+
