+++ backend/src/domain/interfaces/qr-code.interface.ts
@@ -1,0 +1,12 @@
+export interface IQrCodeRepository {
+    generateQRCode(type: string, data: any): Promise<string>;
+    generateQRCodePNG(type: string, data: any): Promise<Buffer>;
+    generateQRCodeSVG(type: string, data: any): Promise<string>;
+    generateQRCodePDF(type: string, data: any): Promise<Buffer>;
+  }
+
+  export interface QRCodeDTO {
+    type: string;
+    data: any;
+  }
+
