--- backend/src/qr-code/qr-code.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { QrCodeService } from './qr-code.service';
-
-describe('QrCodeService', () => {
-  let service: QrCodeService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [QrCodeService],
-    }).compile();
-
-    service = module.get<QrCodeService>(QrCodeService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
