+++ backend/src/shared/constants/qr-code.constants.ts
@@ -1,0 +1,18 @@
+export const QR_CODE_TYPES = {
+    LINK: 'Link',
+    EMAIL: 'Email',
+    TEXT: 'Text',
+    CALL: 'Call',
+    SMS: 'Sms',
+    WHATSAPP: 'Whatsapp',
+    WIFI: 'Wifi',
+    EVENT: 'Event',
+    VCARD: 'VCard',
+    SOCIAL_MEDIA: 'Social Media',
+    APP: 'App',
+    VIDEO: 'Video',
+    FACEBOOK: 'Facebook',
+    TWITTER: 'Twitter',
+    LOCATION: 'Location',
+  };
+
